rename SyncDayTimeRunnable to SyncDaylightCycleRunnable set doDaylightCycle false on plugin enable
