+++ src/main/java/dev/boooiil/historia/core/HistoriaCore.kt
@@ -33,12 +33,13 @@ import dev.boooiil.historia.core.items.events.player.PlayerToggleSprintListener
 import dev.boooiil.historia.core.proficiency.ProficiencyRegistryLoader
 import dev.boooiil.historia.core.proficiency.skills.ISkill
 import dev.boooiil.historia.core.proficiency.skills.SkillRegistryLoader
-import dev.boooiil.historia.core.runnable.SyncDayCycleRunnable
+import dev.boooiil.historia.core.runnable.SyncDaylightCycleRunnable
 import dev.boooiil.historia.core.runnable.SavePlayerRunnable
 import dev.boooiil.historia.core.runnable.TemperaturePollRunnable
 import dev.boooiil.historia.core.runnable.UpdateScoreboardRunnable
 import dev.boooiil.historia.core.util.CoreLogger
 import org.bukkit.Bukkit
+import org.bukkit.GameRule
 import org.bukkit.NamespacedKey
 import org.bukkit.Server
 import org.bukkit.event.Listener
@@ -128,8 +129,9 @@ open class HistoriaCore : JavaPlugin() {
         registerEvent(ChunkLoadListener())
         // end
 
+        server.worlds[0].setGameRule(GameRule.DO_DAYLIGHT_CYCLE, false) // set false in favour of custom daylight cycle
+        registerRunnable(SyncDaylightCycleRunnable(), 0)
         // registerRunnable(new ClassEnchantsRunnable());
-        registerRunnable(SyncDayCycleRunnable(), 0)
         registerRunnable(UpdateScoreboardRunnable())
         registerRunnable(SavePlayerRunnable(), 6000)
         registerRunnable(TemperaturePollRunnable(), 20L)
